requirements.txt
fastapi
uvicorn
pydantic
aiohttp
asyncio
aiofiles
requests
python-dotenv
pytest